+++ R/ac_get_contact_tags.R
@@ -20,8 +20,8 @@ ac_get_contact_tags <- function(
                   add_headers("Api-Token" = Sys.getenv('ACTIVECAMPAGN_API_TOKEN')))
       },
       until =  ~ status_code(.) == 200,
-      interval  = getOption('ractivecampaig.max_tries'),
-      max_tries = getOption('ractivecampaig.interval')
+      interval  = getOption('ractivecampaig.interval'),
+      max_tries = getOption('ractivecampaig.max_tries')
     )
 
     data <- content(ans)