+++ R/ac_get_contacts.R
@@ -97,9 +97,14 @@ ac_get_contacts <- function(
     }
 
     # send request
-    ans <- GET(str_glue("{Sys.getenv('ACTIVECAMPAGN_API_URL')}/api/3/contacts"),
+    retry(
+      {ans <- GET(str_glue("{Sys.getenv('ACTIVECAMPAGN_API_URL')}/api/3/contacts"),
                query = qry,
-               add_headers("Api-Token" = Sys.getenv('ACTIVECAMPAGN_API_TOKEN')))
+               add_headers("Api-Token" = Sys.getenv('ACTIVECAMPAGN_API_TOKEN')))},
+      until =  ~ status_code(.) == 200,
+      interval  = getOption('ractivecampaig.interval'),
+      max_tries = getOption('ractivecampaig.max_tries')
+    )
 
     data <- content(ans)