+++ R/ac_get_custom_deal_fields_values.R
@@ -29,11 +29,18 @@ ac_get_custom_deal_fields_values <- function(
   while ( (is.na(total) | offset <= total) | is_first_iteration  ) {
 
     # send request
-    ans <- GET(str_glue("{Sys.getenv('ACTIVECAMPAGN_API_URL')}/api/3/dealCustomFieldData"),
-               query = list(limit  = limit,
-                            offset = offset,
-                            "filters[dealId]" = deal_id),
-               add_headers("Api-Token" = Sys.getenv('ACTIVECAMPAGN_API_TOKEN')))
+    retry(
+      {
+      ans <- GET(str_glue("{Sys.getenv('ACTIVECAMPAGN_API_URL')}/api/3/dealCustomFieldData"),
+                 query = list(limit  = limit,
+                              offset = offset,
+                              "filters[dealId]" = deal_id),
+                 add_headers("Api-Token" = Sys.getenv('ACTIVECAMPAGN_API_TOKEN')))
+      },
+      until = ~ status_code(.) == 200,
+      interval  = getOption('ractivecampaig.interval'),
+      max_tries = getOption('ractivecampaig.max_tries')
+    )
 
     data <- content(ans)