+++ R/ac_get_deal_piplines.R
@@ -40,8 +40,8 @@ ac_get_deal_piplines <- function(
                  add_headers("Api-Token" = Sys.getenv('ACTIVECAMPAGN_API_TOKEN')))
         },
       until = ~ status_code(.) == 200,
-      interval  = getOption('ractivecampaig.max_tries'),
-      max_tries = getOption('ractivecampaig.interval')
+      interval  = getOption('ractivecampaig.interval'),
+      max_tries = getOption('ractivecampaig.max_tries')
     )