+++ R/ac_get_deals.R
@@ -97,8 +97,8 @@ ac_get_deals <- function(
                  add_headers("Api-Token" = Sys.getenv('ACTIVECAMPAGN_API_TOKEN')))
       },
       until = ~ status_code(.) == 200,
-      interval  = getOption('ractivecampaig.max_tries'),
-      max_tries = getOption('ractivecampaig.interval')
+      interval  = getOption('ractivecampaig.interval'),
+      max_tries = getOption('ractivecampaig.max_tries')
     )
 
     data <- content(ans)