+++ R/ac_get_messages.R
@@ -34,8 +34,8 @@ ac_get_messages <- function(
                    add_headers("Api-Token" = Sys.getenv('ACTIVECAMPAGN_API_TOKEN')))
       },
       until = ~ status_code(.) == 200,
-      interval  = getOption('ractivecampaig.max_tries'),
-      max_tries = getOption('ractivecampaig.interval')
+      interval  = getOption('ractivecampaig.interval'),
+      max_tries = getOption('ractivecampaig.max_tries')
   )
 
     data <- content(ans)